Restructuring gridadmin2geopackage code
