+++ processing/grid_creation_algorithm.py
@@ -10,7 +10,7 @@ from qgis.core import (
     QgsVectorLayer,
 )
 from threedigrid_builder import make_gridadmin, SchematisationError
-from ThreeDiToolbox.processing.results_analysis.utils import gridadmin2geopackage, load_computational_layers
+from ThreeDiToolbox.processing.processing_utils import gridadmin2geopackage, load_computational_layers
 
 
 class ThreeDiGenerateCompGridAlgorithm(QgsProcessingAlgorithm):
