+++ processing/styles/node.qml
