+++ processing/styles/obstacle.qml
